# openapi/integration-api.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Сервис ремонта оборудования — интеграционный API",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "/api/v1",
      "description": "Интеграционный API, версия 1"
    }
  ],
  "paths": {
...
    "/requests": {
      "get": {
        "tags": [
          "requests"
        ],
        "summary": "List Requests",
        "operationId": "list_requests_requests_get",
        "security": [
          {
            "bearer": []
          }
        ],
        "parameters": [
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "array",
                  "items": {
                    "type": "string"
                  }
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          },
          {
            "name": "assignment_state",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "array",
                  "items": {
                    "type": "string"
                  }
                },
                {
                  "type": "null"
                }
              ],
              "title": "Assignment State"
            }
          },
          {
            "name": "updated_since",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date-time"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Updated Since"
            }
          },
          {
            "name": "cursor",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Cursor"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 50,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Page_RequestListItemView_"
                }
              }
            }
          },
          "400": {
            "description": "Некорректный запрос.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Не пройдена аутентификация.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Недостаточно прав или scope.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Объект не найден или недоступен вызывающему.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "Конфликт состояния: версия, идемпотентность или недопустимый переход.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Ошибка проверки полей.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "429": {
            "description": "Слишком много запросов.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "503": {
            "description": "Сервис временно недоступен, повторите позже.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "default": {
            "description": "Ошибка. Формат одинаковый для всех методов.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "x-required-scopes": [
          "requests:read"
        ],
        "description": "Требует scope `requests:read`."
      }
    },
    "/requests/{request_id}": {
      "get": {
        "tags": [
          "requests"
        ],
        "summary": "Get Request",
        "operationId": "get_request_requests__request_id__get",
        "security": [
          {
            "bearer": []
          }
        ],
        "parameters": [
          {
            "name": "request_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Request Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "anyOf": [
                    {
                      "$ref": "#/components/schemas/RequestProviderView"
                    },
                    {
                      "$ref": "#/components/schemas/RequestFormerProviderView"
                    }
                  ],
                  "title": "Response Get Request Requests  Request Id  Get"
                }
              }
            }
          },
          "400": {
            "description": "Некорректный запрос.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Не пройдена аутентификация.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Недостаточно прав или scope.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Объект не найден или недоступен вызывающему.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "Конфликт состояния: версия, идемпотентность или недопустимый переход.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Ошибка проверки полей.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "429": {
            "description": "Слишком много запросов.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "503": {
            "description": "Сервис временно недоступен, повторите позже.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "default": {
            "description": "Ошибка. Формат одинаковый для всех методов.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "x-required-scopes": [
          "requests:read"
        ],
        "description": "Требует scope `requests:read`."
      }
    },
    "/requests/{request_id}/external-reference": {
      "post": {
        "tags": [
          "requests"
        ],
        "summary": "Set External Reference",
        "operationId": "set_external_reference_requests__request_id__external_reference_post",
        "security": [
          {
            "bearer": []
          }
        ],
        "parameters": [
          {
            "name": "request_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Request Id"
            }
          },
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Idempotency-Key"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ExternalReferenceBody"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RequestProviderView"
                }
              }
            }
          },
          "400": {
            "description": "Некорректный запрос.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Не пройдена аутентификация.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Недостаточно прав или scope.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Объект не найден или недоступен вызывающему.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "Конфликт состояния: версия, идемпотентность или недопустимый переход.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Ошибка проверки полей.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "429": {
            "description": "Слишком много запросов.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "503": {
            "description": "Сервис временно недоступен, повторите позже.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "default": {
            "description": "Ошибка. Формат одинаковый для всех методов.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "x-required-scopes": [
          "requests:write"
        ],
        "description": "Требует scope `requests:write`."
      }
    },
    "/requests/{request_id}/accept": {
      "post": {
        "tags": [
          "requests"
        ],
        "summary": "Accept Request",
        "description": "Подтверждение назначения, в том числе выбранного на бирже (offer.select) — по scope заявок, а не биржи.\n\nТребует scope `requests:write`.",
        "operationId": "accept_request_requests__request_id__accept_post",
        "security": [
          {
            "bearer": []
          }
        ],
        "parameters": [
          {
            "name": "request_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Request Id"
            }
          },
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Idempotency-Key"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AcceptRequestBody"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RequestProviderView"
                }
              }
            }
          },
          "400": {
            "description": "Некорректный запрос.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Не пройдена аутентификация.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Недостаточно прав или scope.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Объект не найден или недоступен вызывающему.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "Конфликт состояния: версия, идемпотентность или недопустимый переход.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Ошибка проверки полей.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "429": {
            "description": "Слишком много запросов.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "503": {
            "description": "Сервис временно недоступен, повторите позже.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "default": {
            "description": "Ошибка. Формат одинаковый для всех методов.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "x-required-scopes": [
          "requests:write"
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
      "AcceptRequestBody": {
        "properties": {
          "assignment_id": {
            "type": "string",
            "title": "Assignment Id"
          },
          "expected_version": {
            "type": "integer",
            "minimum": 1.0,
            "title": "Expected Version",
            "description": "Версия заявки, которую видел клиент (`version` карточки)."
          }
        },
        "type": "object",
        "required": [
          "assignment_id",
          "expected_version"
        ],
        "title": "AcceptRequestBody"
      },
...
      "ErrorResponse": {
        "type": "object",
        "required": [
          "error"
        ],
        "properties": {
          "error": {
            "type": "object",
            "required": [
              "code",
              "message",
              "request_id",
              "details"
            ],
            "properties": {
              "code": {
                "type": "string",
                "description": "Машиночитаемый код ошибки, например NOT_FOUND."
              },
              "message": {
                "type": "string",
                "description": "Сообщение на русском для лога/UI."
              },
              "request_id": {
                "type": "string",
                "description": "Идентификатор запроса — тот же, что в заголовке X-Request-ID."
              },
              "details": {
                "type": "object",
                "description": "Дополнительные поля ошибки, зависят от кода.",
                "additionalProperties": true
              }
            }
          }
        }
      },
...
      "ExternalReferenceBody": {
        "properties": {
          "external_id": {
            "type": "string",
            "maxLength": 200,
            "minLength": 1,
            "title": "External Id"
          },
          "expected_version": {
            "type": "integer",
            "minimum": 1.0,
            "title": "Expected Version",
            "description": "Версия заявки, которую видел клиент (`version` карточки)."
          }
        },
        "type": "object",
        "required": [
          "external_id",
          "expected_version"
        ],
        "title": "ExternalReferenceBody"
      },
...
      "Page_RequestListItemView_": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/RequestListItemView"
            },
            "type": "array",
            "title": "Items"
          },
          "next_cursor": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Next Cursor"
          }
        },
        "type": "object",
        "required": [
          "items"
        ],
        "title": "Page[RequestListItemView]"
      },
...
      "RequestFormerProviderView": {
        "properties": {
          "request_id": {
            "type": "string",
            "title": "Request Id"
          },
          "assignment": {
            "$ref": "#/components/schemas/AssignmentView"
          }
        },
        "type": "object",
        "required": [
          "request_id",
          "assignment"
        ],
        "title": "RequestFormerProviderView"
      },
...
    },
    "securitySchemes": {
      "bearer": {
        "type": "http",
        "description": "Ключ интеграции сервисной компании: `Authorization: Bearer <ключ>`.",
        "scheme": "bearer",
        "bearerFormat": "rk_<env>_<prefix>_<secret>"
      }
    }
  }
}